test: validating a use case
